+++ lib/hub/context.rb
@@ -10,10 +10,14 @@ module Hub
 
     # Parses URLs for git remotes and stores info
     REMOTES = Hash.new do |cache, remote|
-      url = GIT_CONFIG["config remote.#{remote}.url"]
+      if remote
+        url = GIT_CONFIG["config remote.#{remote}.url"]
 
-      if url && url.to_s =~ %r{\bgithub\.com[:/](.+)/(.+).git$}
-        cache[remote] = { :user => $1, :repo => $2 }
+        if url && url.to_s =~ %r{\bgithub\.com[:/](.+)/(.+).git$}
+          cache[remote] = { :user => $1, :repo => $2 }
+        else
+          cache[remote] = { }
+        end
       else
         cache[remote] = { }
       end
@@ -61,7 +65,7 @@ module Hub
     end
 
     def remotes
-      list = GIT_CONFIG['remote'].split("\n")
+      list = GIT_CONFIG['remote'].to_s.split("\n")
       main = list.delete('origin') and list.unshift(main)
       list
     end
@@ -71,7 +75,13 @@ module Hub
     end
 
     def current_remote
-      (current_branch && remote_for(current_branch)) || default_remote
+      return if remotes.empty?
+
+      if current_branch
+        remote_for(current_branch)
+      else
+        default_remote
+      end
     end
 
     def default_remote
@@ -92,6 +102,14 @@ module Hub
 
     def http_clone?
       GIT_CONFIG['config --bool hub.http-clone'] == 'true'
+    end
+
+    # Core.repositoryformatversion should exist for all git
+    # repositories, and be blank for all non-git repositories. If
+    # there's a better config setting to check here, this can be
+    # changed without breaking anything.
+    def is_repo?
+      GIT_CONFIG['config core.repositoryformatversion']
     end
 
     def github_url(options = {})