+++ HISTORY.md
@@ -1,3 +1,17 @@
+## 1.4.1 (2010-08-08)
+## 1.4.0 (2010-08-08)
+
+* Added new `hub create` command
+* Added support for `remote set-url`
+* Bugfix: Don't try multiple git commands on a non-git dir when grabbing remote
+* Bugfix: Adding remotes when no remotes exist
+
+## 1.3.2 (2010-07-24)
+
+* bugfix: cherry picking of commit URL
+* bugfix: git init -g
+
+## 1.3.1 (2010-04-29)
 ## 1.3.0 (2010-04-29)
 
 * Tracking branches awareness